Refactor routes API: move GPX generation to server and update DTO flow

+++ src/main/java/com/canyoncompanion/canyon_api/model/entities/WaypointEntity.java
@@ -37,7 +37,7 @@ public class WaypointEntity {
     @Column(nullable = false)
     private Double longitude;
 
-    private Double elevation;
+    private Float elevation;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
